python_da/requirements.txt
pandas==1.3.*
pytest==7.0.*
numpy==1.21.*
scipy==1.7.*
azure-storage-blob==12.9.*
Flask==2.0.*
Werkzeug==2.0.*
azure-core>=1.24.0,<2.0.0
matplotlib==3.4.*
seaborn==0.11.*
pytest-mock==3.*
azure-servicebus==7.*
typing-extensions>=4.6.0
